Remove rtps reader layer
