+++ dds/src/implementation/dds/dds_data_reader.rs
@@ -1,4 +1,7 @@
-use std::collections::{HashMap, HashSet};
+use std::{
+    collections::{HashMap, HashSet},
+    convert::TryInto,
+};
 
 use crate::{
     builtin_topics::{BuiltInTopicKey, PublicationBuiltinTopicData, SubscriptionBuiltinTopicData},
@@ -7,18 +10,26 @@ use crate::{
             discovered_reader_data::{DiscoveredReaderData, ReaderProxy},
             discovered_writer_data::DiscoveredWriterData,
         },
+        data_representation_inline_qos::{
+            parameter_id_values::{PID_KEY_HASH, PID_STATUS_INFO},
+            types::StatusInfo,
+        },
         dds::nodes::DataReaderNode,
         rtps::{
             messages::{
                 overall_structure::{RtpsMessageHeader, RtpsMessageRead, RtpsSubmessageReadKind},
+                submessage_elements::{Data, Parameter, ParameterList},
                 submessages::{
                     data::DataSubmessageRead, data_frag::DataFragSubmessageRead,
                     gap::GapSubmessageRead, heartbeat::HeartbeatSubmessageRead,
                     heartbeat_frag::HeartbeatFragSubmessageRead,
                 },
             },
-            reader::{convert_data_frag_to_cache_change, RtpsReader, RtpsReaderError},
-            types::{Guid, GuidPrefix, Locator, SequenceNumber, ENTITYID_UNKNOWN, GUID_UNKNOWN},
+            reader::RtpsReader,
+            types::{
+                ChangeKind, Guid, GuidPrefix, Locator, SequenceNumber, ENTITYID_UNKNOWN,
+                GUID_UNKNOWN,
+            },
             writer_proxy::RtpsWriterProxy,
         },
         rtps_udp_psm::udp_transport::UdpTransportWrite,
@@ -32,7 +43,8 @@ use crate::{
         instance::InstanceHandle,
         qos::{DataReaderQos, SubscriberQos, TopicQos},
         qos_policy::{
-            DurabilityQosPolicyKind, QosPolicyId, ReliabilityQosPolicyKind, DEADLINE_QOS_POLICY_ID,
+            DestinationOrderQosPolicyKind, DurabilityQosPolicyKind, HistoryQosPolicyKind,
+            QosPolicyId, ReliabilityQosPolicyKind, DEADLINE_QOS_POLICY_ID,
             DESTINATIONORDER_QOS_POLICY_ID, DURABILITY_QOS_POLICY_ID, LATENCYBUDGET_QOS_POLICY_ID,
             LIVELINESS_QOS_POLICY_ID, PRESENTATION_QOS_POLICY_ID, RELIABILITY_QOS_POLICY_ID,
         },
@@ -45,9 +57,9 @@ use crate::{
     },
     subscription::{
         data_reader::Sample,
-        sample_info::{InstanceStateKind, SampleStateKind, ViewStateKind},
+        sample_info::{InstanceStateKind, SampleInfo, SampleStateKind, ViewStateKind},
     },
-    topic_definition::type_support::{DdsDeserialize, DdsType},
+    topic_definition::type_support::{dds_deserialize, DdsDeserialize, DdsSerializedKey, DdsType},
 };
 
 use super::{
@@ -55,6 +67,176 @@ use super::{
     dds_subscriber::DdsSubscriber, message_receiver::MessageReceiver, nodes::SubscriberNode,
     status_condition_impl::StatusConditionImpl,
 };
+
+pub type RtpsReaderResult<T> = Result<T, RtpsReaderError>;
+
+#[derive(Debug, PartialEq, Eq)]
+pub enum RtpsReaderError {
+    InvalidData(&'static str),
+    Rejected(InstanceHandle, SampleRejectedStatusKind),
+}
+
+pub struct RtpsReaderCacheChange {
+    kind: ChangeKind,
+    writer_guid: Guid,
+    data: Data,
+    inline_qos: ParameterList,
+    source_timestamp: Option<Time>,
+    sample_state: SampleStateKind,
+    disposed_generation_count: i32,
+    no_writers_generation_count: i32,
+    reception_timestamp: Time,
+}
+
+pub fn convert_data_frag_to_cache_change(
+    data_frag_submessage: &DataFragSubmessageRead,
+    data: Vec<u8>,
+    source_timestamp: Option<Time>,
+    source_guid_prefix: GuidPrefix,
+    reception_timestamp: Time,
+) -> Result<RtpsReaderCacheChange, RtpsReaderError> {
+    let writer_guid = Guid::new(source_guid_prefix, data_frag_submessage.writer_id());
+
+    let inline_qos = data_frag_submessage.inline_qos();
+
+    let change_kind = if data_frag_submessage.key_flag() {
+        if let Some(p) = inline_qos
+            .parameter()
+            .iter()
+            .find(|&x| x.parameter_id() == PID_STATUS_INFO)
+        {
+            let mut deserializer =
+                cdr::Deserializer::<_, _, cdr::LittleEndian>::new(p.value(), cdr::Infinite);
+            let status_info: StatusInfo =
+                serde::Deserialize::deserialize(&mut deserializer).unwrap();
+            match status_info {
+                STATUS_INFO_DISPOSED => Ok(ChangeKind::NotAliveDisposed),
+                STATUS_INFO_UNREGISTERED => Ok(ChangeKind::NotAliveUnregistered),
+                STATUS_INFO_DISPOSED_UNREGISTERED => Ok(ChangeKind::NotAliveDisposedUnregistered),
+                _ => Err(RtpsReaderError::InvalidData("Unknown status info value")),
+            }
+        } else {
+            Err(RtpsReaderError::InvalidData(
+                "Missing mandatory StatusInfo parameter",
+            ))
+        }
+    } else {
+        Ok(ChangeKind::Alive)
+    }?;
+
+    Ok(RtpsReaderCacheChange {
+        kind: change_kind,
+        writer_guid,
+        data: Data::new(data),
+        inline_qos,
+        source_timestamp,
+        sample_state: SampleStateKind::NotRead,
+        disposed_generation_count: 0, // To be filled up only when getting stored
+        no_writers_generation_count: 0, // To be filled up only when getting stored
+        reception_timestamp,
+    })
+}
+
+struct InstanceHandleBuilder(fn(&mut &[u8]) -> RtpsReaderResult<DdsSerializedKey>);
+
+impl InstanceHandleBuilder {
+    fn new<Foo>() -> Self
+    where
+        Foo: for<'de> serde::Deserialize<'de> + DdsType,
+    {
+        fn deserialize_data_to_key<Foo>(data: &mut &[u8]) -> RtpsReaderResult<DdsSerializedKey>
+        where
+            Foo: for<'de> serde::Deserialize<'de> + DdsType,
+        {
+            Ok(dds_deserialize::<Foo>(data)
+                .map_err(|_| RtpsReaderError::InvalidData("Failed to deserialize data"))?
+                .get_serialized_key())
+        }
+
+        Self(deserialize_data_to_key::<Foo>)
+    }
+
+    fn build_instance_handle(
+        &self,
+        change_kind: ChangeKind,
+        mut data: &[u8],
+        inline_qos: &[Parameter],
+    ) -> RtpsReaderResult<InstanceHandle> {
+        Ok(match change_kind {
+            ChangeKind::Alive | ChangeKind::AliveFiltered => (self.0)(&mut data)?.into(),
+            ChangeKind::NotAliveDisposed
+            | ChangeKind::NotAliveUnregistered
+            | ChangeKind::NotAliveDisposedUnregistered => match inline_qos
+                .iter()
+                .find(|&x| x.parameter_id() == PID_KEY_HASH)
+            {
+                Some(p) => InstanceHandle::new(p.value().try_into().unwrap()),
+                None => dds_deserialize::<DdsSerializedKey>(data)
+                    .map_err(|_| RtpsReaderError::InvalidData("Failed to deserialize key"))?
+                    .into(),
+            },
+        })
+    }
+}
+
+struct Instance {
+    view_state: ViewStateKind,
+    instance_state: InstanceStateKind,
+    most_recent_disposed_generation_count: i32,
+    most_recent_no_writers_generation_count: i32,
+}
+
+impl Instance {
+    fn new() -> Self {
+        Self {
+            view_state: ViewStateKind::New,
+            instance_state: InstanceStateKind::Alive,
+            most_recent_disposed_generation_count: 0,
+            most_recent_no_writers_generation_count: 0,
+        }
+    }
+
+    fn update_state(&mut self, change_kind: ChangeKind) {
+        match self.instance_state {
+            InstanceStateKind::Alive => {
+                if change_kind == ChangeKind::NotAliveDisposed
+                    || change_kind == ChangeKind::NotAliveDisposedUnregistered
+                {
+                    self.instance_state = InstanceStateKind::NotAliveDisposed;
+                } else if change_kind == ChangeKind::NotAliveUnregistered {
+                    self.instance_state = InstanceStateKind::NotAliveNoWriters;
+                }
+            }
+            InstanceStateKind::NotAliveDisposed => {
+                if change_kind == ChangeKind::Alive {
+                    self.instance_state = InstanceStateKind::Alive;
+                    self.most_recent_disposed_generation_count += 1;
+                }
+            }
+            InstanceStateKind::NotAliveNoWriters => {
+                if change_kind == ChangeKind::Alive {
+                    self.instance_state = InstanceStateKind::Alive;
+                    self.most_recent_no_writers_generation_count += 1;
+                }
+            }
+        }
+
+        match self.view_state {
+            ViewStateKind::New => (),
+            ViewStateKind::NotNew => {
+                if change_kind == ChangeKind::NotAliveDisposed
+                    || change_kind == ChangeKind::NotAliveUnregistered
+                {
+                    self.view_state = ViewStateKind::New;
+                }
+            }
+        }
+    }
+
+    fn mark_viewed(&mut self) {
+        self.view_state = ViewStateKind::NotNew;
+    }
+}
 
 pub enum UserDefinedReaderDataSubmessageReceivedResult {
     NoChange,
@@ -195,6 +377,10 @@ impl SubscriptionMatchedStatus {
 pub struct DdsDataReader {
     rtps_reader: RtpsReader,
     matched_writers: Vec<RtpsWriterProxy>,
+    changes: Vec<RtpsReaderCacheChange>,
+    qos: DataReaderQos,
+    instance_handle_builder: InstanceHandleBuilder,
+    instances: HashMap<InstanceHandle, Instance>,
     type_name: &'static str,
     topic_name: String,
     liveliness_changed_status: LivelinessChangedStatus,
@@ -214,13 +400,19 @@ pub struct DdsDataReader {
 }
 
 impl DdsDataReader {
-    pub fn new(
+    pub fn new<Foo>(
         rtps_reader: RtpsReader,
         type_name: &'static str,
         topic_name: String,
+        qos: DataReaderQos,
         listener: Option<Actor<DdsDataReaderListener>>,
         status_kind: Vec<StatusKind>,
-    ) -> Self {
+    ) -> Self
+    where
+        Foo: DdsType + for<'de> serde::Deserialize<'de>,
+    {
+        let instance_handle_builder = InstanceHandleBuilder::new::<Foo>();
+
         DdsDataReader {
             rtps_reader,
             matched_writers: Vec::new(),
@@ -240,6 +432,10 @@ impl DdsDataReader {
             status_condition: DdsShared::new(DdsRwLock::new(StatusConditionImpl::default())),
             status_kind,
             listener,
+            changes: Vec::new(),
+            qos,
+            instance_handle_builder,
+            instances: HashMap::new(),
         }
     }
 
@@ -429,10 +625,10 @@ impl DdsDataReader {
             .find(|wp| wp.remote_writer_guid() == writer_guid)
         {
             let expected_seq_num = writer_proxy.available_changes_max() + 1;
-            match self.rtps_reader.get_qos().reliability.kind {
+            match self.qos.reliability.kind {
                 ReliabilityQosPolicyKind::BestEffort => {
                     if sequence_number >= expected_seq_num {
-                        let change_result = self.rtps_reader.convert_data_to_cache_change(
+                        let change_result = self.convert_data_to_cache_change(
                             data_submessage,
                             source_timestamp,
                             source_guid_prefix,
@@ -440,7 +636,7 @@ impl DdsDataReader {
                         );
                         match change_result {
                             Ok(change) => {
-                                let add_change_result = self.rtps_reader.add_change(change);
+                                let add_change_result = self.add_change(change);
 
                                 match add_change_result {
                                     Ok(instance_handle) => {
@@ -465,7 +661,7 @@ impl DdsDataReader {
                 }
                 ReliabilityQosPolicyKind::Reliable => {
                     if sequence_number == expected_seq_num {
-                        let change_result = self.rtps_reader.convert_data_to_cache_change(
+                        let change_result = self.convert_data_to_cache_change(
                             data_submessage,
                             source_timestamp,
                             source_guid_prefix,
@@ -473,7 +669,7 @@ impl DdsDataReader {
                         );
                         match change_result {
                             Ok(change) => {
-                                let add_change_result = self.rtps_reader.add_change(change);
+                                let add_change_result = self.add_change(change);
 
                                 match add_change_result {
                                     Ok(instance_handle) => {
@@ -492,7 +688,7 @@ impl DdsDataReader {
                 }
             }
         } else if data_submessage.reader_id() == ENTITYID_UNKNOWN {
-            let change_result = self.rtps_reader.convert_data_to_cache_change(
+            let change_result = self.convert_data_to_cache_change(
                 data_submessage,
                 source_timestamp,
                 source_guid_prefix,
@@ -500,7 +696,7 @@ impl DdsDataReader {
             );
             match change_result {
                 Ok(change) => {
-                    let add_change_result = self.rtps_reader.add_change(change);
+                    let add_change_result = self.add_change(change);
 
                     match add_change_result {
                         Ok(h) => DataReceivedResult::NewSampleAdded(h),
@@ -591,7 +787,7 @@ impl DdsDataReader {
             .find(|wp| wp.remote_writer_guid() == writer_guid)
         {
             let expected_seq_num = writer_proxy.available_changes_max() + 1;
-            match self.rtps_reader.get_qos().reliability.kind {
+            match self.qos.reliability.kind {
                 ReliabilityQosPolicyKind::BestEffort => {
                     if sequence_number >= expected_seq_num {
                         writer_proxy.push_data_frag(data_frag_submessage);
@@ -605,7 +801,7 @@ impl DdsDataReader {
                             );
                             match change_results {
                                 Ok(change) => {
-                                    let add_change_result = self.rtps_reader.add_change(change);
+                                    let add_change_result = self.add_change(change);
                                     match add_change_result {
                                         Ok(instance_handle) => {
                                             writer_proxy.received_change_set(sequence_number);
@@ -646,7 +842,7 @@ impl DdsDataReader {
 
                             match change_result {
                                 Ok(change) => {
-                                    let add_change_result = self.rtps_reader.add_change(change);
+                                    let add_change_result = self.add_change(change);
                                     match add_change_result {
                                         Ok(instance_handle) => {
                                             writer_proxy.received_change_set(sequence_number);
@@ -726,7 +922,7 @@ impl DdsDataReader {
         heartbeat_submessage: &HeartbeatSubmessageRead,
         source_guid_prefix: GuidPrefix,
     ) {
-        if self.rtps_reader.get_qos().reliability.kind == ReliabilityQosPolicyKind::Reliable {
+        if self.qos.reliability.kind == ReliabilityQosPolicyKind::Reliable {
             let writer_guid = Guid::new(source_guid_prefix, heartbeat_submessage.writer_id());
 
             if let Some(writer_proxy) = self
@@ -758,7 +954,7 @@ impl DdsDataReader {
         heartbeat_frag_submessage: &HeartbeatFragSubmessageRead,
         source_guid_prefix: GuidPrefix,
     ) {
-        if self.rtps_reader.get_qos().reliability.kind == ReliabilityQosPolicyKind::Reliable {
+        if self.qos.reliability.kind == ReliabilityQosPolicyKind::Reliable {
             let writer_guid = Guid::new(source_guid_prefix, heartbeat_frag_submessage.writer_id());
 
             if let Some(writer_proxy) = self
@@ -872,7 +1068,6 @@ impl DdsDataReader {
         subscriber_qos: &SubscriberQos,
     ) -> Vec<QosPolicyId> {
         let writer_info = discovered_writer_data.dds_publication_data();
-        let reader_qos = self.rtps_reader.get_qos().clone();
 
         let mut incompatible_qos_policy_list = Vec::new();
 
@@ -884,22 +1079,22 @@ impl DdsDataReader {
         {
             incompatible_qos_policy_list.push(PRESENTATION_QOS_POLICY_ID);
         }
-        if &reader_qos.durability > writer_info.durability() {
+        if &self.qos.durability > writer_info.durability() {
             incompatible_qos_policy_list.push(DURABILITY_QOS_POLICY_ID);
         }
-        if &reader_qos.deadline > writer_info.deadline() {
+        if &self.qos.deadline > writer_info.deadline() {
             incompatible_qos_policy_list.push(DEADLINE_QOS_POLICY_ID);
         }
-        if &reader_qos.latency_budget > writer_info.latency_budget() {
+        if &self.qos.latency_budget > writer_info.latency_budget() {
             incompatible_qos_policy_list.push(LATENCYBUDGET_QOS_POLICY_ID);
         }
-        if &reader_qos.liveliness > writer_info.liveliness() {
+        if &self.qos.liveliness > writer_info.liveliness() {
             incompatible_qos_policy_list.push(LIVELINESS_QOS_POLICY_ID);
         }
-        if reader_qos.reliability.kind > writer_info.reliability().kind {
+        if self.qos.reliability.kind > writer_info.reliability().kind {
             incompatible_qos_policy_list.push(RELIABILITY_QOS_POLICY_ID);
         }
-        if &reader_qos.destination_order > writer_info.destination_order() {
+        if &self.qos.destination_order > writer_info.destination_order() {
             incompatible_qos_policy_list.push(DESTINATIONORDER_QOS_POLICY_ID);
         }
 
@@ -943,13 +1138,28 @@ impl DdsDataReader {
             return Err(DdsError::NotEnabled);
         }
 
-        self.rtps_reader.read(
+        self.status_condition
+            .write_lock()
+            .remove_communication_state(StatusKind::DataAvailable);
+
+        let indexed_sample_list = self.create_indexed_sample_collection::<Foo>(
             max_samples,
             sample_states,
             view_states,
             instance_states,
             specific_instance_handle,
-        )
+        )?;
+
+        let change_index_list: Vec<usize>;
+        let samples;
+
+        (change_index_list, samples) = indexed_sample_list.into_iter().map(|(i, s)| (i, s)).unzip();
+
+        for index in change_index_list {
+            self.changes[index].sample_state = SampleStateKind::Read;
+        }
+
+        Ok(samples)
     }
 
     pub fn take<Foo>(
@@ -967,13 +1177,28 @@ impl DdsDataReader {
             return Err(DdsError::NotEnabled);
         }
 
-        self.rtps_reader.take(
+        let indexed_sample_list = self.create_indexed_sample_collection::<Foo>(
             max_samples,
             sample_states,
             view_states,
             instance_states,
             specific_instance_handle,
-        )
+        )?;
+
+        self.status_condition
+            .write_lock()
+            .remove_communication_state(StatusKind::DataAvailable);
+
+        let mut change_index_list: Vec<usize>;
+        let samples;
+
+        (change_index_list, samples) = indexed_sample_list.into_iter().map(|(i, s)| (i, s)).unzip();
+
+        while let Some(index) = change_index_list.pop() {
+            self.changes.remove(index);
+        }
+
+        Ok(samples)
     }
 
     pub fn read_next_instance<Foo>(
@@ -991,13 +1216,16 @@ impl DdsDataReader {
             return Err(DdsError::NotEnabled);
         }
 
-        self.rtps_reader.read_next_instance(
-            max_samples,
-            previous_handle,
-            sample_states,
-            view_states,
-            instance_states,
-        )
+        match self.next_instance(previous_handle) {
+            Some(next_handle) => self.read(
+                max_samples,
+                sample_states,
+                view_states,
+                instance_states,
+                Some(next_handle),
+            ),
+            None => Err(DdsError::NoData),
+        }
     }
 
     pub fn take_next_instance<Foo>(
@@ -1015,13 +1243,16 @@ impl DdsDataReader {
             return Err(DdsError::NotEnabled);
         }
 
-        self.rtps_reader.take_next_instance(
-            max_samples,
-            previous_handle,
-            sample_states,
-            view_states,
-            instance_states,
-        )
+        match self.next_instance(previous_handle) {
+            Some(next_handle) => self.take(
+                max_samples,
+                sample_states,
+                view_states,
+                instance_states,
+                Some(next_handle),
+            ),
+            None => Err(DdsError::NoData),
+        }
     }
 
     pub fn get_key_value<Foo>(
@@ -1047,7 +1278,7 @@ impl DdsDataReader {
             Ok(())
         }?;
 
-        match self.rtps_reader.get_qos().durability.kind {
+        match self.qos.durability.kind {
             DurabilityQosPolicyKind::Volatile => Err(DdsError::IllegalOperation),
             DurabilityQosPolicyKind::TransientLocal => Ok(()),
         }?;
@@ -1060,16 +1291,17 @@ impl DdsDataReader {
 
     pub fn set_qos(&mut self, qos: DataReaderQos) -> DdsResult<()> {
         if self.is_enabled() {
-            self.rtps_reader.get_qos().check_immutability(&qos)?;
+            self.qos.check_immutability(&qos)?;
         }
 
-        self.rtps_reader.set_qos(qos)?;
+        qos.is_consistent()?;
+        self.qos = qos;
 
         Ok(())
     }
 
     pub fn get_qos(&self) -> DataReaderQos {
-        self.rtps_reader.get_qos().clone()
+        self.qos.clone()
     }
 
     pub fn get_instance_handle(&self) -> InstanceHandle {
@@ -1084,7 +1316,6 @@ impl DdsDataReader {
         default_multicast_locator_list: Vec<Locator>,
     ) -> DiscoveredReaderData {
         let guid = self.rtps_reader.guid();
-        let reader_qos = self.rtps_reader.get_qos().clone();
 
         let unicast_locator_list = if self.rtps_reader.unicast_locator_list().is_empty() {
             default_unicast_locator_list
@@ -1113,15 +1344,15 @@ impl DdsDataReader {
                 },
                 self.topic_name.clone(),
                 self.type_name.to_string(),
-                reader_qos.durability.clone(),
-                reader_qos.deadline.clone(),
-                reader_qos.latency_budget.clone(),
-                reader_qos.liveliness.clone(),
-                reader_qos.reliability.clone(),
-                reader_qos.ownership.clone(),
-                reader_qos.destination_order.clone(),
-                reader_qos.user_data.clone(),
-                reader_qos.time_based_filter,
+                self.qos.durability.clone(),
+                self.qos.deadline.clone(),
+                self.qos.latency_budget.clone(),
+                self.qos.liveliness.clone(),
+                self.qos.reliability.clone(),
+                self.qos.ownership.clone(),
+                self.qos.destination_order.clone(),
+                self.qos.user_data.clone(),
+                self.qos.time_based_filter,
                 subscriber_qos.presentation.clone(),
                 subscriber_qos.partition.clone(),
                 topic_qos.topic_data,
@@ -1151,8 +1382,7 @@ impl DdsDataReader {
             .instance_reception_time
             .iter()
             .partition(|&(_, received_time)| {
-                DurationKind::Finite(now - *received_time)
-                    > self.rtps_reader.get_qos().deadline.period
+                DurationKind::Finite(now - *received_time) > self.qos.deadline.period
             });
 
         self.instance_reception_time = instance_reception_time;
@@ -1437,5 +1667,396 @@ impl DdsDataReader {
 
     pub fn guid(&self) -> Guid {
         self.rtps_reader.guid()
+    }
+
+    pub fn convert_data_to_cache_change(
+        &self,
+        data_submessage: &DataSubmessageRead,
+        source_timestamp: Option<Time>,
+        source_guid_prefix: GuidPrefix,
+        reception_timestamp: Time,
+    ) -> RtpsReaderResult<RtpsReaderCacheChange> {
+        let writer_guid = Guid::new(source_guid_prefix, data_submessage.writer_id());
+
+        let data = data_submessage.serialized_payload();
+
+        let inline_qos = data_submessage.inline_qos();
+
+        let change_kind = match (data_submessage.data_flag(), data_submessage.key_flag()) {
+            (true, false) => Ok(ChangeKind::Alive),
+            (false, true) | (false, false) => {
+                if let Some(p) = inline_qos
+                    .parameter()
+                    .iter()
+                    .find(|&x| x.parameter_id() == PID_STATUS_INFO)
+                {
+                    let mut deserializer =
+                        cdr::Deserializer::<_, _, cdr::LittleEndian>::new(p.value(), cdr::Infinite);
+                    let status_info: StatusInfo =
+                        serde::Deserialize::deserialize(&mut deserializer).unwrap();
+                    match status_info {
+                        STATUS_INFO_DISPOSED => Ok(ChangeKind::NotAliveDisposed),
+                        STATUS_INFO_UNREGISTERED => Ok(ChangeKind::NotAliveUnregistered),
+                        STATUS_INFO_DISPOSED_UNREGISTERED => {
+                            Ok(ChangeKind::NotAliveDisposedUnregistered)
+                        }
+                        _ => Err(RtpsReaderError::InvalidData("Unknown status info value")),
+                    }
+                } else {
+                    Err(RtpsReaderError::InvalidData(
+                        "Missing mandatory StatusInfo parameter",
+                    ))
+                }
+            }
+            (true, true) => Err(RtpsReaderError::InvalidData(
+                "Invalid data and key flag combination",
+            )),
+        }?;
+
+        Ok(RtpsReaderCacheChange {
+            kind: change_kind,
+            writer_guid,
+            data,
+            inline_qos,
+            source_timestamp,
+            sample_state: SampleStateKind::NotRead,
+            disposed_generation_count: 0, // To be filled up only when getting stored
+            no_writers_generation_count: 0, // To be filled up only when getting stored
+            reception_timestamp,
+        })
+    }
+
+    fn is_max_samples_limit_reached(&self, change_instance_handle: &InstanceHandle) -> bool {
+        let total_samples = self
+            .changes
+            .iter()
+            .filter(|cc| {
+                &self
+                    .instance_handle_builder
+                    .build_instance_handle(cc.kind, cc.data.as_ref(), cc.inline_qos.parameter())
+                    .expect("Change in cache must have valid instance handle")
+                    == change_instance_handle
+            })
+            .count();
+
+        total_samples == self.qos.resource_limits.max_samples
+    }
+
+    fn is_max_instances_limit_reached(&self, change_instance_handle: &InstanceHandle) -> bool {
+        let instance_handle_list: HashSet<_> = self
+            .changes
+            .iter()
+            .map(|cc| {
+                self.instance_handle_builder
+                    .build_instance_handle(cc.kind, cc.data.as_ref(), cc.inline_qos.parameter())
+                    .expect("Change in cache must have valid instance handle")
+            })
+            .collect();
+
+        if instance_handle_list.contains(change_instance_handle) {
+            false
+        } else {
+            instance_handle_list.len() == self.qos.resource_limits.max_instances
+        }
+    }
+
+    fn is_max_samples_per_instance_limit_reached(
+        &self,
+        change_instance_handle: &InstanceHandle,
+    ) -> bool {
+        let total_samples_of_instance = self
+            .changes
+            .iter()
+            .filter(|cc| {
+                &self
+                    .instance_handle_builder
+                    .build_instance_handle(cc.kind, cc.data.as_ref(), cc.inline_qos.parameter())
+                    .expect("Change in cache must have valid instance handle")
+                    == change_instance_handle
+            })
+            .count();
+
+        total_samples_of_instance == self.qos.resource_limits.max_samples_per_instance
+    }
+
+    fn is_sample_of_interest_based_on_time(
+        &self,
+        change: &RtpsReaderCacheChange,
+        change_instance_handle: &InstanceHandle,
+    ) -> bool {
+        let closest_timestamp_before_received_sample = self
+            .changes
+            .iter()
+            .filter(|cc| {
+                &self
+                    .instance_handle_builder
+                    .build_instance_handle(cc.kind, cc.data.as_ref(), cc.inline_qos.parameter())
+                    .expect("Change in cache must have valid instance handle")
+                    == change_instance_handle
+            })
+            .filter(|cc| cc.source_timestamp <= change.source_timestamp)
+            .map(|cc| cc.source_timestamp)
+            .max();
+
+        if let Some(Some(t)) = closest_timestamp_before_received_sample {
+            if let Some(sample_source_time) = change.source_timestamp {
+                let sample_separation = sample_source_time - t;
+                DurationKind::Finite(sample_separation)
+                    >= self.qos.time_based_filter.minimum_separation
+            } else {
+                true
+            }
+        } else {
+            true
+        }
+    }
+
+    pub fn add_change(
+        &mut self,
+        mut change: RtpsReaderCacheChange,
+    ) -> RtpsReaderResult<InstanceHandle> {
+        let change_instance_handle = self.instance_handle_builder.build_instance_handle(
+            change.kind,
+            change.data.as_ref(),
+            change.inline_qos.parameter(),
+        )?;
+        if self.is_sample_of_interest_based_on_time(&change, &change_instance_handle) {
+            if self.is_max_samples_limit_reached(&change_instance_handle) {
+                Err(RtpsReaderError::Rejected(
+                    change_instance_handle,
+                    SampleRejectedStatusKind::RejectedBySamplesLimit,
+                ))
+            } else if self.is_max_instances_limit_reached(&change_instance_handle) {
+                Err(RtpsReaderError::Rejected(
+                    change_instance_handle,
+                    SampleRejectedStatusKind::RejectedByInstancesLimit,
+                ))
+            } else if self.is_max_samples_per_instance_limit_reached(&change_instance_handle) {
+                Err(RtpsReaderError::Rejected(
+                    change_instance_handle,
+                    SampleRejectedStatusKind::RejectedBySamplesPerInstanceLimit,
+                ))
+            } else {
+                let num_alive_samples_of_instance = self
+                    .changes
+                    .iter()
+                    .filter(|cc| {
+                        self.instance_handle_builder
+                            .build_instance_handle(
+                                cc.kind,
+                                cc.data.as_ref(),
+                                cc.inline_qos.parameter(),
+                            )
+                            .unwrap()
+                            == change_instance_handle
+                            && cc.kind == ChangeKind::Alive
+                    })
+                    .count() as i32;
+
+                if let HistoryQosPolicyKind::KeepLast(depth) = self.qos.history.kind {
+                    if depth == num_alive_samples_of_instance {
+                        let index_sample_to_remove = self
+                            .changes
+                            .iter()
+                            .position(|cc| {
+                                self.instance_handle_builder
+                                    .build_instance_handle(
+                                        cc.kind,
+                                        cc.data.as_ref(),
+                                        cc.inline_qos.parameter(),
+                                    )
+                                    .unwrap()
+                                    == change_instance_handle
+                                    && cc.kind == ChangeKind::Alive
+                            })
+                            .expect("Samples must exist");
+                        self.changes.remove(index_sample_to_remove);
+                    }
+                }
+
+                let instance_entry = self
+                    .instances
+                    .entry(change_instance_handle)
+                    .or_insert_with(Instance::new);
+
+                instance_entry.update_state(change.kind);
+
+                change.disposed_generation_count =
+                    instance_entry.most_recent_disposed_generation_count;
+                change.no_writers_generation_count =
+                    instance_entry.most_recent_no_writers_generation_count;
+                self.changes.push(change);
+
+                match self.qos.destination_order.kind {
+                    DestinationOrderQosPolicyKind::BySourceTimestamp => {
+                        self.changes.sort_by(|a, b| {
+                            a.source_timestamp
+                                .as_ref()
+                                .expect("Missing source timestamp")
+                                .cmp(
+                                    b.source_timestamp
+                                        .as_ref()
+                                        .expect("Missing source timestamp"),
+                                )
+                        });
+                    }
+                    DestinationOrderQosPolicyKind::ByReceptionTimestamp => self
+                        .changes
+                        .sort_by(|a, b| a.reception_timestamp.cmp(&b.reception_timestamp)),
+                }
+
+                Ok(change_instance_handle)
+            }
+        } else {
+            Ok(change_instance_handle)
+        }
+    }
+
+    fn create_indexed_sample_collection<Foo>(
+        &mut self,
+        max_samples: i32,
+        sample_states: &[SampleStateKind],
+        view_states: &[ViewStateKind],
+        instance_states: &[InstanceStateKind],
+        specific_instance_handle: Option<InstanceHandle>,
+    ) -> DdsResult<Vec<(usize, Sample<Foo>)>>
+    where
+        Foo: for<'de> DdsDeserialize<'de>,
+    {
+        if let Some(h) = specific_instance_handle {
+            if !self.instances.contains_key(&h) {
+                return Err(DdsError::BadParameter);
+            }
+        };
+
+        let mut indexed_samples = Vec::new();
+
+        let instance_handle_build = &self.instance_handle_builder;
+        let instances = &self.instances;
+        let mut instances_in_collection = HashMap::new();
+        for (index, cache_change) in self
+            .changes
+            .iter_mut()
+            .enumerate()
+            .filter(|(_, cc)| {
+                let sample_instance_handle = instance_handle_build
+                    .build_instance_handle(cc.kind, cc.data.as_ref(), cc.inline_qos.parameter())
+                    .unwrap();
+
+                sample_states.contains(&cc.sample_state)
+                    && view_states.contains(&instances[&sample_instance_handle].view_state)
+                    && instance_states.contains(&instances[&sample_instance_handle].instance_state)
+                    && if let Some(h) = specific_instance_handle {
+                        h == sample_instance_handle
+                    } else {
+                        true
+                    }
+            })
+            .take(max_samples as usize)
+        {
+            let sample_instance_handle = self
+                .instance_handle_builder
+                .build_instance_handle(
+                    cache_change.kind,
+                    cache_change.data.as_ref(),
+                    cache_change.inline_qos.parameter(),
+                )
+                .unwrap();
+            instances_in_collection
+                .entry(sample_instance_handle)
+                .or_insert_with(Instance::new);
+
+            instances_in_collection
+                .get_mut(&sample_instance_handle)
+                .unwrap()
+                .update_state(cache_change.kind);
+            let sample_state = cache_change.sample_state;
+            let view_state = self.instances[&sample_instance_handle].view_state;
+            let instance_state = self.instances[&sample_instance_handle].instance_state;
+
+            let absolute_generation_rank = (self.instances[&sample_instance_handle]
+                .most_recent_disposed_generation_count
+                + self.instances[&sample_instance_handle].most_recent_no_writers_generation_count)
+                - (instances_in_collection[&sample_instance_handle]
+                    .most_recent_disposed_generation_count
+                    + instances_in_collection[&sample_instance_handle]
+                        .most_recent_no_writers_generation_count);
+
+            let (data, valid_data) = match cache_change.kind {
+                ChangeKind::Alive | ChangeKind::AliveFiltered => (
+                    Some(
+                        dds_deserialize(cache_change.data.as_ref())
+                            .map_err(|_err| DdsError::Error)?,
+                    ),
+                    true,
+                ),
+                ChangeKind::NotAliveDisposed
+                | ChangeKind::NotAliveUnregistered
+                | ChangeKind::NotAliveDisposedUnregistered => (None, false),
+            };
+
+            let sample_info = SampleInfo {
+                sample_state,
+                view_state,
+                instance_state,
+                disposed_generation_count: cache_change.disposed_generation_count,
+                no_writers_generation_count: cache_change.no_writers_generation_count,
+                sample_rank: 0,     // To be filled up after collection is created
+                generation_rank: 0, // To be filled up after collection is created
+                absolute_generation_rank,
+                source_timestamp: cache_change.source_timestamp,
+                instance_handle: sample_instance_handle,
+                publication_handle: cache_change.writer_guid.into(),
+                valid_data,
+            };
+
+            let sample = Sample { data, sample_info };
+
+            indexed_samples.push((index, sample))
+        }
+
+        // After the collection is created, update the relative generation rank values and mark the read instances as viewed
+        for handle in instances_in_collection.into_keys() {
+            let most_recent_sample_absolute_generation_rank = indexed_samples
+                .iter()
+                .filter(|(_, s)| s.sample_info.instance_handle == handle)
+                .map(|(_, s)| s.sample_info.absolute_generation_rank)
+                .last()
+                .expect("Instance handle must exist on collection");
+
+            let mut total_instance_samples_in_collection = indexed_samples
+                .iter()
+                .filter(|(_, s)| s.sample_info.instance_handle == handle)
+                .count();
+
+            for (_, sample) in indexed_samples
+                .iter_mut()
+                .filter(|(_, s)| s.sample_info.instance_handle == handle)
+            {
+                sample.sample_info.generation_rank = sample.sample_info.absolute_generation_rank
+                    - most_recent_sample_absolute_generation_rank;
+
+                total_instance_samples_in_collection -= 1;
+                sample.sample_info.sample_rank = total_instance_samples_in_collection as i32;
+            }
+
+            self.instances
+                .get_mut(&handle)
+                .expect("Sample must exist on hash map")
+                .mark_viewed()
+        }
+
+        if indexed_samples.is_empty() {
+            Err(DdsError::NoData)
+        } else {
+            Ok(indexed_samples)
+        }
+    }
+
+    fn next_instance(&self, previous_handle: Option<InstanceHandle>) -> Option<InstanceHandle> {
+        match previous_handle {
+            Some(p) => self.instances.keys().filter(|&h| h > &p).min().cloned(),
+            None => self.instances.keys().min().cloned(),
+        }
     }
 }
